+++ MiniProject/Assets/01.Script/Data/PlayerData.cs
@@ -4,18 +4,18 @@ using UnityEngine;
 
 public class PlayerData : TableBase
 {
-	public int level;
-	public int healthPoint;
-	public int damage;
-	public int moveSpeed;
-	public int criticalChange;
-	public int criticalDamage;
-	public int armor;
+	public float level;
+	public float healthPoint;
+	public float damage;
+	public float moveSpeed;
+	public float criticalChange;
+	public float criticalDamage;
+	public float armor;
 	public PlayerData()
 	{
 
 	}
-	public PlayerData(int l,int h,int d,int mS,int cC,int cD,int a)
+	public PlayerData(float l, float h, float d, float mS, float cC, float cD, float a)
 	{
 		level = l;
 		healthPoint = h;
@@ -25,7 +25,7 @@ public class PlayerData : TableBase
 		criticalDamage = cD;
 		armor = a;
 	}
-	public override int GetTableID()
+	public override float GetTableID()
 	{
 		return level;
 	}
